requirements.txt
langchain
python-dotenv
langchain_community
pypdf
beautifulsoup4
pymupdf
langchain-text-splitters
langchain_openai
chromadb
langchain-ollama
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
streamlit
faiss-cpu
transformers
accelerate